Update polling.yaml
Configuration of "polling.yaml" file is not correct.
Polling.yaml is written same as pipeline .yaml.

Polling.yaml should contain only meters name that will be polled at some fixed interval. There is no need for sink part in polling.yaml, it is required only in pipeline.yaml.
Identation and syntax for meters name is also corrected. (It should not be in quotes)